refactor: Cleanup of structure (#30)
* refactor: Name the probe module and helpers for what they hold

`track_probe::TrackProbeError` said "track" twice for one idea, so the
module is `probe`, its error is `ProbeError` and what it returns is
`Probed`. Inside it `track_codec`, `track_kind` and `track_fragments`
become `codec`, `kind` and `fragments`, reading the way `frame_rate` and
`language` beside them already did — the struct literal is now
`codec: codec(&boxes)?`.

track_helpers.rs folds into track.rs. It only ever held free functions
over `Track`, and its tests carried a second copy of the same kind
constructors the ones in track.rs used. `probe_all_tracks` had the same
stutter and belongs to the type, so it is `Track::probe_all`, sitting
next to the `Track::probe` it calls; the bitrate and duration helpers
stay free functions in the module.


* refactor: Split dyndo-text into a fragmenter and a muxer

wvtt.rs did three jobs: work out where fragments end, tile cues into the
samples covering them, and write boxes. Only the last is about `wvtt`.

fragmenter.rs takes the first two. It divides a subtitle at the asset's
splice points and on the requested grid, tiling each fragment into samples
that borrow the cues on screen over them. A sample holding no cues is just a
sample holding no cues; deciding that means an empty `vtte` is the muxer's
business, since gapless coverage is what ISO/IEC 14496-30 asks for and not
something a subtitle knows about. One `intervals` helper serves both passes,
a fragment being the timeline divided at the asset's clock and a sample a
fragment divided at the cue edges.

muxer/ holds one module per container, so a second slots in beside `wvtt`
the way a second parser would beside VTT. The `vttc`, `payl` and `vtte`
boxes move to muxer/atoms.rs, crate-private beside the only muxer that
writes them: mp4-atom models the `wvtt` sample entry but not the cue boxes
filling a sample, in 0.12.1 or in any release since. Its `VttC` is the
`vttC` configuration box, one capital away from the `vttc` here.

wvtt.rs is left with pack and the box builders. The interval tests follow
the logic they cover and stop encoding a whole mp4 to assert on integers;
what stays behind reads boxes back out.


* refactor: Count milliseconds in u32

A text track's timescale is always 1000, so its times are milliseconds
outright, and the boxes carrying them are u32: a `trun` entry's duration and
a `sidx` reference's subsegment duration both. Carrying them as u64 bought
no range anything can use — u32 runs to 49 days — and left a narrowing at
the muxer, where it was an `expect`.

That expect was reachable. `99999999:00:00.000` is a well-formed VTT
timestamp; it parsed to 3.6e14 ms, sailed through fragmenting, and panicked
in pack — inside the read layer, on every request for that path. Parsing to
u32 rejects it as the malformed timestamp it always was, at the layer that
already has an error for it.

The rest follows to keep one rule: u32 unless something needs u64. Segment
options and the CLI flags behind them are milliseconds, as are Track's four
duration functions. A fragment's raw duration comes straight off a u32 sidx
field, so `u64::from` there was ceremony over a value that already fit.

u64 stays where it earns its keep. Byte offsets and sizes, since a source
passes 4 GiB. Timescale units that accumulate — a segment's raw duration, a
cumulative index, an earliest presentation time — because u32 in those runs
out after 13 hours at 90 kHz, not 49 days. Bitrates, which u32 caps at 4.29
Gbps and which dash-mpd and hls_m3u8 want wide anyway.

Two conversions go: the boundary narrowing inside fragment(), and the
`u64::from` in probe. Three widenings arrive, all infallible, where a
fragment's duration joins a segment's and where DASH asks for a Duration.

Segment options are wire-facing, so an asset.json or query naming a length
above u32 now fails to parse rather than being taken. Forty-nine days of
segment length was never a thing to honour.


---------

+++ crates/dyndo-core/src/segment.rs
@@ -9,16 +9,16 @@ pub struct SegmentOptions {
     /// The shortest a served segment may be, in milliseconds; fragments are
     /// grouped until they reach it.
     #[serde(default, alias = "sml", alias = "segment_min_length")]
-    pub min_length: u64,
+    pub min_length: u32,
     /// How long each segment of a packaged subtitle track is, in milliseconds.
     /// Unlike `min_length` this is exact, since dyndo fragments those tracks
     /// itself rather than grouping what a file already contains. Zero asks for no
     /// grid, leaving the asset's splice points as the only cuts.
     #[serde(default, alias = "stl", alias = "segment_text_length")]
-    pub text_length: u64,
+    pub text_length: u32,
     /// Times a segment has to start at, in milliseconds.
     #[serde(default, alias = "sb", alias = "segment_boundaries")]
-    pub boundaries: Vec<u64>,
+    pub boundaries: Vec<u32>,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -61,7 +61,7 @@ fn group_fragments(
             .map(|fragment| Segment {
                 byte_offset: fragment.byte_offset,
                 byte_size: fragment.byte_size,
-                raw_duration: fragment.raw_duration,
+                raw_duration: u64::from(fragment.raw_duration),
             })
             .collect();
     }
@@ -70,7 +70,7 @@ fn group_fragments(
     let mut cumulative = Vec::with_capacity(fragments.len() + 1);
     cumulative.push(0u64);
     for fragment in fragments {
-        cumulative.push(cumulative[cumulative.len() - 1] + fragment.raw_duration());
+        cumulative.push(cumulative[cumulative.len() - 1] + u64::from(fragment.raw_duration()));
     }
     let cuts = snap_cuts(&cumulative, timescale, &options.boundaries);
 
@@ -99,7 +99,7 @@ fn group_fragments(
     segments
 }
 
-fn snap_cuts(cumulative: &[u64], timescale: u32, boundaries: &[u64]) -> Vec<usize> {
+fn snap_cuts(cumulative: &[u64], timescale: u32, boundaries: &[u32]) -> Vec<usize> {
     let mut cuts: Vec<usize> = boundaries
         .iter()
         .map(|&boundary| {
@@ -133,7 +133,7 @@ mod tests {
         assert_eq!((options.min_length, options.text_length), (0, 0));
     }
 
-    fn options(min_length: u64, boundaries: &[u64]) -> SegmentOptions {
+    fn options(min_length: u32, boundaries: &[u32]) -> SegmentOptions {
         SegmentOptions {
             min_length,
             boundaries: boundaries.to_vec(),
@@ -141,7 +141,7 @@ mod tests {
         }
     }
 
-    fn fragments(raw_durations: &[u64]) -> Vec<Fragment> {
+    fn fragments(raw_durations: &[u32]) -> Vec<Fragment> {
         let mut byte_offset = 100;
         raw_durations
             .iter()
